fix deadlock in RangeScan
Here is a simple (and obvious method) for using RangeScan:

    rowsc, errc = reader.RangeScan()
    for row := range rowsc {
        // ...
    }
    err = <-errc
    if err != nil {
         t.Error("Error running query: ", err)
    }

Previously, this would deadlock if an error occurred because gossie was
blocking on trying to write to the `errc` channel while the user of
gossie is blocked on reading from `rowsc`.

A simple fix for this is just to add a buffer to the errc channel so we
can queue up the error for the caller to read.